refactor(tm-66): split renderer into ES modules
Breaks the 3000-line main.js into 16 focused modules under
src/renderer/modules/:

  state       — shared state + all constants
  utils       — pure utility functions (fmt*, escHtml, etc.)
  store       — saveState / loadState (async electron-store)
  toast       — showToast / showConfirm
  theme       — initTheme / applyTheme
  ripple      — initRipple
  week        — ISO week helpers, buildWeekDays, navigation
  summary     — calcDayTotalMins / updateSummary
  render      — renderAll/renderDays, buildDayCard, drag-and-drop, toggleDay
  entry-modal — full entry CRUD modal + undo delete
  copy-to     — copy-to-week modal
  recurring   — recurring task management
  scheduled   — scheduled task management
  context-menu — right-click menu + quick-view popup
  report      — TXT generation, preview, copy, download, print
  search      — inline + advanced search + navigateToResult
  star        — starred entries
  sidebar     — sidebar/about, auto-updater, keyboard shortcuts

main.js is now a clean ~60-line entry point that imports and
calls each module's init function from DOMContentLoaded.
